#2896 comment: moved SQL queries from project.py to new file sql_queries.py
